refactor: reorganize file picker into platform and API directories
