+++ lib/file_picker.dart
@@ -1,8 +1,8 @@
 export 'src/file_picker.dart';
-export 'src/platform_file.dart';
-export 'src/file_picker_result.dart';
-export 'src/file_picker_types.dart';
-export 'src/file_picker_method_channel.dart';
-export 'src/file_picker_macos.dart';
-export 'src/linux/file_picker_linux.dart';
-export 'src/windows/file_picker_windows.dart';
+export 'src/api/platform_file.dart';
+export 'src/api/file_picker_result.dart';
+export 'src/api/file_picker_types.dart';
+export 'src/platform/file_picker_method_channel.dart';
+export 'src/platform/macos/file_picker_macos.dart';
+export 'src/platform/linux/file_picker_linux.dart';
+export 'src/platform/windows/file_picker_windows.dart';
