+++ lib/src/api/exceptions.dart
