+++ lib/src/platform/linux/file_picker_linux.dart
@@ -1,14 +1,14 @@
 import 'dart:async';
 import 'dart:typed_data';
 
-import 'package:file_picker/src/file_picker_types.dart';
-import 'package:file_picker/src/file_picker_platform_interface.dart';
-import 'package:file_picker/src/file_picker_result.dart';
-import 'package:file_picker/src/platform_file.dart';
+import 'package:file_picker/src/api/file_picker_types.dart';
+import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
+import 'package:file_picker/src/api/file_picker_result.dart';
+import 'package:file_picker/src/api/platform_file.dart';
 import 'package:file_picker/src/utils.dart';
-import 'package:file_picker/src/linux/xdp_filechooser.dart';
-import 'package:file_picker/src/linux/xdp_request.dart';
-import 'package:file_picker/src/linux/filters.dart';
+import 'package:file_picker/src/platform/linux/xdp_filechooser.dart';
+import 'package:file_picker/src/platform/linux/xdp_request.dart';
+import 'package:file_picker/src/platform/linux/filters.dart';
 import 'package:dbus/dbus.dart';
 
 class FilePickerLinux extends FilePickerPlatform {
