+++ lib/src/platform/linux/filters.dart
@@ -1,4 +1,4 @@
-import 'package:file_picker/src/file_picker_types.dart';
+import 'package:file_picker/src/api/file_picker_types.dart';
 import 'package:dbus/dbus.dart';
 
 typedef FilterInfo = Map<String, List<(int, String)>>;
