+++ lib/src/platform/macos/file_picker_macos.dart
@@ -1,7 +1,7 @@
-import 'package:file_picker/src/file_picker_types.dart';
-import 'package:file_picker/src/platform_file.dart';
-import 'package:file_picker/src/file_picker_result.dart';
-import 'package:file_picker/src/file_picker_platform_interface.dart';
+import 'package:file_picker/src/api/file_picker_types.dart';
+import 'package:file_picker/src/api/platform_file.dart';
+import 'package:file_picker/src/api/file_picker_result.dart';
+import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
 import 'package:file_picker/src/utils.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
