+++ lib/src/platform/web/file_picker_web.dart
@@ -2,10 +2,10 @@ import 'dart:async';
 import 'dart:js_interop';
 import 'dart:typed_data';
 
-import 'package:file_picker/src/file_picker_types.dart';
-import 'package:file_picker/src/file_picker_result.dart';
-import 'package:file_picker/src/platform_file.dart';
-import 'package:file_picker/src/file_picker_platform_interface.dart';
+import 'package:file_picker/src/api/file_picker_types.dart';
+import 'package:file_picker/src/api/file_picker_result.dart';
+import 'package:file_picker/src/api/platform_file.dart';
+import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:path/path.dart' as p;
 import 'package:web/web.dart';
