+++ lib/src/platform/windows/file_picker_windows.dart
@@ -6,13 +6,13 @@ import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:ffi/ffi.dart';
-import 'package:file_picker/src/file_picker_types.dart';
-import 'package:file_picker/src/file_picker_result.dart';
+import 'package:file_picker/src/api/file_picker_types.dart';
+import 'package:file_picker/src/api/file_picker_result.dart';
 
-import 'package:file_picker/src/exceptions.dart';
-import 'package:file_picker/src/file_picker_platform_interface.dart';
+import 'package:file_picker/src/api/exceptions.dart';
+import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
 import 'package:file_picker/src/utils.dart';
-import 'package:file_picker/src/windows/file_picker_windows_ffi_types.dart';
+import 'package:file_picker/src/platform/windows/file_picker_windows_ffi_types.dart';
 import 'package:flutter/foundation.dart';
 import 'package:path/path.dart';
 import 'package:win32/win32.dart';
