+++ test/file_picker_windows_test.dart
@@ -4,10 +4,11 @@ library;
 import 'dart:ffi';
 
 import 'package:ffi/ffi.dart';
-import 'package:file_picker/src/exceptions.dart';
-import 'package:file_picker/src/file_picker_types.dart';
-import 'package:file_picker/src/windows/file_picker_windows.dart';
-import 'package:file_picker/src/windows/file_picker_windows_ffi_types.dart';
+import 'package:file_picker/src/api/exceptions.dart';
+import 'package:file_picker/src/api/file_picker_types.dart';
+
+import 'package:file_picker/src/platform/windows/file_picker_windows.dart';
+import 'package:file_picker/src/platform/windows/file_picker_windows_ffi_types.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
