+++ test/linux/xdp_filter_test.dart
@@ -1,8 +1,8 @@
 @TestOn('linux')
 library;
 
-import 'package:file_picker/src/file_picker_types.dart';
-import 'package:file_picker/src/linux/filters.dart';
+import 'package:file_picker/src/api/file_picker_types.dart';
+import 'package:file_picker/src/platform/linux/filters.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
